feat(demodata): Simplify getDemoDataBucketMembership function
